Разобрал на отдельные сборки. Скомпилировал
