+++ FileManager/ConsoleUI/ConsoleUI.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Exeptions\Exeptions.csproj" />
+    <ProjectReference Include="..\Repos\Repos.csproj" />
+    <ProjectReference Include="..\TaskManager\TaskManager.csproj" />
+  </ItemGroup>
+
+</Project>
