+++ FileManager/ConsoleUI/PrintConsoleCLS.cs
