+++ FileManager/ConsoleUI/PrintConsoleCOPY.cs
