+++ FileManager/ConsoleUI/PrintConsoleFILEINF.cs
