+++ FileManager/ConsoleUI/PrintConsoleProcDelete.cs
