+++ FileManager/ConsoleUI/PseudoConsoleUI.cs
@@ -10,7 +10,7 @@ namespace FileManager
     /// </summary>
     public class PseudoConsoleUI
     {
-        #region Fields and properties
+
 
         /// <summary>
         /// Константа запроса на ввод информации
@@ -52,297 +52,121 @@ namespace FileManager
         /// </summary>
         public const int PAGE_LINES = 40;
 
+
+
+        #region For delete
+
+        ///// <summary>
+        ///// Метод выводит на экран скопированные файлы
+        ///// </summary>
+        //public static void ShowCopyCommand(List<string> files)
+        //{
+        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
+
+        //    Console.WriteLine();
+
+        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
+
+        //    Console.WriteLine($"Sucsessfully copied {files.Count / 2} files:");
+
+        //    for (int i = 0; i < files.Count; i += 2)
+        //    {
+        //        Console.SetCursorPosition(0, Console.BufferHeight - 1);
+
+        //        Console.WriteLine($"Copy {files[i]} to {files[i + 1]} is done.");
+        //    }
+        //    Console.WriteLine();
+        //}
+
         #endregion
 
-        /// <summary>
-        /// Метод выводит на экран скопированные файлы
-        /// </summary>
-        public static void ShowCopyCommand(List<string> files)
-        {
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
+        #region For delete
 
-            Console.WriteLine();
 
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
+        ///// <summary>
+        ///// Метод выводит на экран состояние текущей директории
+        ///// </summary>
+        ///// <param name="dirFiles">DrivesAndDirectories актуальное состояние программы</param>
+        //public static void PrintDirectoryProrerties(CurrentDrivesAndDirs dirFiles)
+        //{
+        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
 
-            Console.WriteLine($"Sucsessfully copied {files.Count / 2} files:");
+        //    Console.WriteLine();
 
-            for (int i = 0; i < files.Count; i += 2)
-            {
-                Console.SetCursorPosition(0, Console.BufferHeight - 1);
+        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
 
-                Console.WriteLine($"Copy {files[i]} to {files[i + 1]} is done.");
-            }
-            Console.WriteLine();
-        }
+        //    Console.ForegroundColor = ConsoleColor.DarkGray;
 
-        /// <summary>
-        /// Метод выводит на экран все подкаталоги и файлы текущего каталога
-        /// </summary>
-        /// <param name="dirFiles">DrivesAndDirectories актуальное состояние программы</param>
-        /// <param name="dirStart">int первый каталог на странице</param>
-        /// <param name="dirStop">int крайний каталог на странице</param>
-        /// <param name="fileStart">int первый файл на странице</param>
-        /// <param name="fileStop">int крайний файл на странице</param>
-        public static void PrintAllSubdirectoriesAndFilesByPages(DrivesDirectoriesFilesArray dirFiles, int dirStart, int dirStop, int fileStart, int fileStop)
-        {
-            FileInfo[] files = dirFiles.Files;
+        //    Console.Write($"Текущая директория: ");
 
-            DirectoryInfo[] subdirectories = dirFiles.Directories;
+        //    Console.ForegroundColor = ConsoleColor.Magenta;
 
-            Console.WriteLine();
+        //    Console.Write($"{dirFiles.CuttentDirectory.FullName}");
 
-            for (int i = dirStart; i < dirStop; i++)
-            {
-                Console.SetCursorPosition(0, Console.BufferHeight - 1);
+        //    Console.ForegroundColor = ConsoleColor.DarkGray;
 
-                Console.Write($"{subdirectories[i].Name}");
+        //    Console.Write($" содержит: ");
 
-                Console.SetCursorPosition(20, Console.BufferHeight - 1);
+        //    Console.ForegroundColor = ConsoleColor.Magenta;
 
-                Console.ForegroundColor = ConsoleColor.DarkGray;
+        //    Console.Write($"{dirFiles.CuttentDirectory.GetFiles().Length}");
 
-                Console.Write($"<DIR>");
+        //    Console.ForegroundColor = ConsoleColor.DarkGray;
 
-                Console.ForegroundColor = ConsoleColor.White;
+        //    Console.Write($" файлов,");
 
-                Console.SetCursorPosition(30, Console.BufferHeight - 1);
+        //    Console.ForegroundColor = ConsoleColor.Magenta;
 
-                Console.Write($"{subdirectories[i].CreationTime:dd-MM-yyyy}");
+        //    Console.WriteLine();
 
-                Console.WriteLine();
-            }
-            for (int i = fileStart; i < fileStop; i++)
-            {
-                Console.ForegroundColor = ConsoleColor.DarkYellow;
+        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
 
-                Console.SetCursorPosition(0, Console.BufferHeight - 1);
+        //    Console.ForegroundColor = ConsoleColor.DarkGray;
 
-                Console.Write($"{ Path.GetFileNameWithoutExtension(files[i].FullName)}");
+        //    Console.Write($"имеет атрибуты: ");
 
-                Console.SetCursorPosition(25, Console.BufferHeight - 1);
+        //    Console.ForegroundColor = ConsoleColor.Magenta;
 
-                Console.ForegroundColor = ConsoleColor.DarkGreen;
+        //    Console.Write($"{dirFiles.CuttentDirectory.Attributes}");
 
-                Console.Write($"{ Path.GetExtension(files[i].FullName)}");
+        //    Console.ForegroundColor = ConsoleColor.DarkGray;
 
-                Console.ForegroundColor = ConsoleColor.White;
+        //    Console.Write($", Время создания: ");
 
-                Console.SetCursorPosition(30, Console.BufferHeight - 1);
+        //    Console.ForegroundColor = ConsoleColor.Magenta;
 
-                Console.Write($"{files[i].Length} Bytes");
+        //    Console.Write($"{dirFiles.CuttentDirectory.CreationTime:dd-MM-yyyy HH-mm-ss}");
 
-                Console.SetCursorPosition(70, Console.BufferHeight - 1);
+        //    Console.ForegroundColor = ConsoleColor.White;
 
-                Console.Write($"{ files[i].CreationTime:HH-mm-ss dd-MM-yyyy}");
+        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
 
-                Console.SetCursorPosition(90, Console.BufferHeight - 1);
+        //    Console.WriteLine();
+        //}
 
-                Console.ForegroundColor = ConsoleColor.DarkGreen;
+        #endregion
 
-                Console.Write($"{ files[i].Attributes}");
+        #region For delete
 
-                Console.ForegroundColor = ConsoleColor.White;
+        ///// <summary>
+        ///// Метод выводит на экран Help - лист
+        ///// </summary>
+        //public static void WriteHelp(List<string> list)
+        //{
+        //    Console.WriteLine();
 
-                Console.WriteLine();
-            }
-        }
+        //    foreach (string row in list)
+        //    {
+        //        Console.SetCursorPosition(0, Console.BufferHeight - 1);
 
-        /// <summary>
-        /// Метод выводит в консоль все логические диски
-        /// </summary>
-        /// <param name="drives">DriveInfo[] массив логических дисков</param>
-        public static void PrintAllDrives(DriveInfo[] drives)
-        {
-            string r;
+        //        Console.WriteLine(row);
+        //    }
+        //    Console.WriteLine();
 
-            Console.WriteLine();
+        //}
 
-            for (int i = 0; i < drives.Length; i++)
-            {
-                r = drives[i].IsReady ? "READY" : "NOT READY";
+        #endregion
 
-                Console.ForegroundColor = ConsoleColor.White;
-
-                Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                Console.Write($"{drives[i].Name}");
-
-                Console.SetCursorPosition(4, Console.BufferHeight - 1);
-
-                Console.Write($"{r}");
-
-                Console.SetCursorPosition(10, Console.BufferHeight - 1);
-
-                Console.Write($"{drives[i].DriveType}");
-
-                Console.SetCursorPosition(20, Console.BufferHeight - 1);
-
-                Console.Write($"{drives[i].DriveFormat}");
-
-                Console.SetCursorPosition(27, Console.BufferHeight - 1);
-
-                Console.Write($"Доступно {((drives[i].TotalFreeSpace / 1024) / 1024)} Мбайт");
-
-                Console.SetCursorPosition(60, Console.BufferHeight - 1);
-
-                Console.Write($"из {((drives[i].TotalSize / 1024) / 1024)} Мбайт");
-
-                Console.WriteLine();
-            }
-        }
-
-        /// <summary>
-        /// Метод выводит на экран текущий номер страницы
-        /// </summary>
-        /// <param name="allLines">int количество всех подкаталогов и файлов в каталоге</param>
-        /// <param name="userPage">int страница</param>
-        public static void PrintPageNumber(int allLines, int userPage)
-        {
-            if (userPage != -1)
-            {
-                Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                int p = userPage > (1 + allLines / PAGE_LINES) ? (1 + allLines / PAGE_LINES) : userPage;
-
-                Console.Write($"Страница {p} из {(1 + allLines / PAGE_LINES)}");
-            }
-        }
-
-        /// <summary>
-        /// Метод выводит в консоль параметры полученного файла
-        /// </summary>
-        /// <param name="file">FileInfo файл, параметры которого нужно изучить</param>
-        public static void PrintFileProperties(FileInfo file)
-        {
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.WriteLine();
-
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($"Информация по файлу: ");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.Write($"{file.FullName}");
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($" размер: ");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.Write($"{file.Length} Байт");
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.WriteLine();
-
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($"имеет атрибуты: ");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.Write($"{file.Attributes}");
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($", Время создания: ");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.Write($"{file.CreationTime:dd-MM-yyyy HH-mm-ss}");
-
-            Console.ForegroundColor = ConsoleColor.White;
-
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.WriteLine();
-        }
-
-        /// <summary>
-        /// Метод выводит на экран состояние текущей директории
-        /// </summary>
-        /// <param name="dirFiles">DrivesAndDirectories актуальное состояние программы</param>
-        public static void PrintDirectoryProrerties(DrivesAndDirectories dirFiles)
-        {
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.WriteLine();
-
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($"Текущая директория: ");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.Write($"{dirFiles.CuttentDirectory.FullName}");
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($" содержит: ");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.Write($"{dirFiles.CuttentDirectory.GetFiles().Length}");
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($" файлов,");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.WriteLine();
-
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($"имеет атрибуты: ");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.Write($"{dirFiles.CuttentDirectory.Attributes}");
-
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-
-            Console.Write($", Время создания: ");
-
-            Console.ForegroundColor = ConsoleColor.Magenta;
-
-            Console.Write($"{dirFiles.CuttentDirectory.CreationTime:dd-MM-yyyy HH-mm-ss}");
-
-            Console.ForegroundColor = ConsoleColor.White;
-
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.WriteLine();
-        }
-
-        /// <summary>
-        /// Метод выводит на экран Help - лист
-        /// </summary>
-        public static void WriteHelp(List<string> list)
-        {
-            Console.WriteLine();
-
-            foreach (string row in list)
-            {
-                Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                Console.WriteLine(row);
-            }
-            Console.WriteLine();
-
-        }
 
         /// <summary>
         /// Метод задает размеры окна и буфера консоли
@@ -360,7 +184,7 @@ namespace FileManager
         /// метод перемещает курсор на указанную позицию
         /// </summary>
         /// /// <param name="dir">DrivesAndDirectories актуальное состояние программы</param>
-        public static void SetCursorPosition(DrivesAndDirectories temp)
+        public static void SetCursorPosition(CurrentDrivesAndDirs temp)
         {
             Console.ForegroundColor = ConsoleColor.White;
 
@@ -369,22 +193,12 @@ namespace FileManager
             Console.Write(temp.CuttentDirectory + INPUT_REQUEST);
         }
 
-        /// <summary>
-        /// Метод перещает курсор на указанную позицию для работы с процессами
-        /// </summary>
-        private static void SetCursorPositionForProcesses()
-        {
-            Console.ForegroundColor = ConsoleColor.White;
-
-            Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-            Console.Write(INPUT_REQUEST);
-        }
+        
 
         /// <summary>
         /// Метод запрашивает сведения о процессе
         /// </summary>
-        private static void RequestToEnterProcess(string proc)
+        public static void RequestToEnterProcess(string proc)
         {
             Console.ForegroundColor = ConsoleColor.White;
 
@@ -399,150 +213,22 @@ namespace FileManager
             SetCursorPositionForProcesses();
         }
 
-        /// <summary>
-        /// Метод удаляет процесс
-        /// </summary>
-        public static void DeleteProcess()
-        {
-            BasicLogic logic = new BasicLogic();
+        
 
-            int userDataInt = 0;
-
-            bool isItId;
-
-            RequestToEnterProcess(@"Введите имя процесса или его ID,");
-
-            string userData = Console.ReadLine();
-
-            if (userData != "")
-            {
-                try
-                {
-                    userDataInt = Convert.ToInt32(userData);
-
-                    isItId = true;
-                }
-                catch (Exception)
-                {
-                    isItId = false;
-                }
-
-                if (isItId)
-                {
-                    isItId = logic.KillProcess(userDataInt);
-                }
-                else
-                {
-                    isItId = logic.KillProcess(userData);
-                }
-                if (isItId)
-                {
-                    Console.ForegroundColor = ConsoleColor.White;
-
-                    Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                    Console.WriteLine("Process is sussessfally terminated...Press any key");
-
-                    Console.ReadKey();
-                }
-                else
-                {
-                    Console.ForegroundColor = ConsoleColor.White;
-
-                    Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                    Console.WriteLine("Process not found...Press any key");
-
-                    Console.ReadKey();
-                }
-            }
-        }
 
         /// <summary>
-        /// Метод запускает процесс
+        /// Метод перещает курсор на указанную позицию для работы с процессами
         /// </summary>
-        public static void CreateProcess()
+        private static void SetCursorPositionForProcesses()
         {
-            BasicLogic logic = new BasicLogic();
+            Console.ForegroundColor = ConsoleColor.White;
 
-            bool isItId;
+            Console.SetCursorPosition(0, Console.BufferHeight - 1);
 
-            RequestToEnterProcess(@"Введите имя файла или URL для запуска процесса,");
-
-            string userData = Console.ReadLine();
-
-            if (userData != "")
-            {
-                isItId = logic.CreateNewProcess(userData);
-
-                if (!isItId)
-                {
-                    Console.ForegroundColor = ConsoleColor.White;
-
-                    Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                    Console.WriteLine($"The file {userData} is not exist. Press any key");
-
-                    Console.ReadKey();
-                }
-                else
-                {
-                    Console.ForegroundColor = ConsoleColor.White;
-
-                    Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                    Console.WriteLine($"Process {userData} is launched. Press any key");
-
-                    Console.ReadKey();
-                }
-            }
+            Console.Write(PseudoConsoleUI.INPUT_REQUEST);
         }
 
-        /// <summary>
-        /// Метод выводит в консоль все процессы
-        /// </summary>
-        /// <param name="process"></param>
-        public static void PrintProcesses(Process[] process)
-        {
-            Console.Clear();
 
-            int pgNum = 1;
-
-            for (int i = 0; i < process.Length; i++)
-            {
-                string str = "";
-
-                int stringLength = process[i].ProcessName.Length + Convert.ToString(process[i].Id).Length;
-
-                if (stringLength < LENGTH_LINE)
-                {
-                    for (int j = 0; j < LENGTH_LINE - stringLength; j++)
-                    {
-                        str = str + ".";
-                    }
-                }
-                Console.ForegroundColor = ConsoleColor.White;
-
-                Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                Console.WriteLine($" {String.Format("{0:d4}", i)} {process[i].ProcessName}{str}{process[i].Id}");
-
-                if (i == pgNum * (PAGE_LINES - 1))
-                {
-                    Console.ForegroundColor = ConsoleColor.DarkGreen;
-
-                    Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                    Console.WriteLine($"Страница {pgNum}. Для продолжения нажмите любую клавишу...");
-
-                    Console.ReadKey();
-
-                    Console.ForegroundColor = ConsoleColor.White;
-
-                    pgNum++;
-                }
-            }
-        }
