+++ FileManager/Core/Core.csproj
@@ -1,7 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
@@ -10,7 +9,10 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Core\Core.csproj" />
+    <ProjectReference Include="..\ConsoleUI\ConsoleUI.csproj" />
+    <ProjectReference Include="..\Exeptions\Exeptions.csproj" />
+    <ProjectReference Include="..\LogicDirectives\LogicDirectives.csproj" />
+    <ProjectReference Include="..\Repos\Repos.csproj" />
   </ItemGroup>
 
 </Project>
