+++ FileManager/LogicDirectives/DirectiveCD.cs
