+++ FileManager/LogicDirectives/DirectiveCLS.cs
