+++ FileManager/LogicDirectives/DirectiveCOPY.cs
