+++ FileManager/LogicDirectives/DirectiveDEL.cs
