+++ FileManager/LogicDirectives/DirectiveDELTREE.cs
