+++ FileManager/LogicDirectives/DirectiveFILEINF.cs
