+++ FileManager/LogicDirectives/DirectiveMOVE.cs
