+++ FileManager/TaskManager/BasicLogic.cs
