+++ FileManager/TaskManager/TaskManager.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Exeptions\Exeptions.csproj" />
+  </ItemGroup>
+
+</Project>
